requirements.txt
python-dotenv
numpy
termcolor
rich
